finlay-backend/requirements.txt
fastapi
uvicorn
groq
hindsight-sdk
websockets
pydantic
python-dotenv
yfinance
